Exposed leeway setting in the JWT library
If the system clock of the document server is even a second slower than the
nextcloud server, the JWT token will be rejected as being "not yet valid". This
can be seen in the logs (Nextcloud -> Settings -> Administration -> Logging).

The error message shown to the user when this happens is more generic:

"Error when trying to connect (Error occurred in the document service: Error
while downloading the document file to be converted.)"

This error message has been reported on the bug tracker several times, but it is
not clear when the connection failure was due to imperfectly synchronized clocks
and when it was due to network connectivity, DNS issues, the JWT not being sent,
the secret being wrong, or some other issue.

Here are the tickets that may be fixed by this:
https://github.com/ONLYOFFICE/onlyoffice-nextcloud/issues/345
https://github.com/ONLYOFFICE/onlyoffice-nextcloud/issues/548
https://github.com/ONLYOFFICE/onlyoffice-nextcloud/issues/315
https://github.com/ONLYOFFICE/onlyoffice-nextcloud/issues/290
https://github.com/ONLYOFFICE/onlyoffice-nextcloud/issues/127

There's also one on the owncloud fork that may be fixed by this as well:
https://github.com/ONLYOFFICE/onlyoffice-owncloud/issues/278